pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["Eugene B <ebondarikk@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
async-timeout = "4.0.2"
attrs = "23.1.0"
basehash = "3.0.10"
build = "0.10.0"
cachecontrol = "0.12.11"
cachetools = "5.3.0"
certifi = "2022.12.7"
cffi = "1.15.1"
charset-normalizer = "3.0.1"
cleo = "2.0.1"
crashtest = "0.4.1"
cryptography = "40.0.1"
distlib = "0.3.6"
dulwich = "0.21.3"
filelock = "3.12.0"
firebase = "4.0.1"
firebase-admin = "6.1.0"
google-api-core = "2.11.0"
google-api-python-client = "2.84.0"
google-auth = "2.17.2"
google-auth-httplib2 = "0.1.0"
google-cloud-core = "2.3.2"
google-cloud-firestore = "2.11.0"
google-cloud-storage = "2.8.0"
google-crc32c = "1.5.0"
google-resumable-media = "2.4.1"
googleapis-common-protos = "1.59.0"
grpcio = "1.53.0"
grpcio-status = "1.53.0"
hiredis = "2.2.2"
html5lib = "1.1"
httplib2 = "0.22.0"
idna = "3.4"
importlib-metadata = "6.5.0"
installer = "0.7.0"
jaraco-classes = "3.2.3"
jeepney = "0.8.0"
jsonschema = "4.17.3"
keyring = "23.13.1"
lockfile = "0.12.2"
more-itertools = "9.1.0"
msgpack = "1.0.5"
packaging = "23.1"
pexpect = "4.8.0"
pkginfo = "1.9.6"
platformdirs = "2.6.2"
poetry = "1.4.2"
poetry-core = "1.5.2"
poetry-plugin-export = "1.3.1"
proto-plus = "1.22.2"
protobuf = "4.22.1"
ptyprocess = "0.7.0"
pyasn1 = "0.4.8"
pyasn1-modules = "0.2.8"
pycparser = "2.21"
pyjwt = "2.6.0"
pyparsing = "3.0.9"
pyproject-hooks = "1.0.0"
pyrsistent = "0.19.3"
pytelegrambotapi = "4.10.0"
python-decouple = "3.8"
rapidfuzz = "2.15.1"
redis = "4.5.4"
requests = "2.28.2"
requests-toolbelt = "0.10.1"
rsa = "4.9"
secretstorage = "3.3.3"
shellingham = "1.5.0.post1"
six = "1.16.0"
tomli = "2.0.1"
tomlkit = "0.11.7"
trove-classifiers = "2023.4.18"
uritemplate = "4.1.1"
urllib3 = "1.26.14"
virtualenv = "20.21.0"
webencodings = "0.5.1"
zipp = "3.15.0"
aiohttp = "^3.8.4"
aioredis = "^2.0.1"
docker = "^6.1.0"
google-cloud = "^0.34.0"
gcloud = "^0.18.3"
google-cloud-api-keys = "^0.5.2"
google-cloud-functions = "^1.12.0"
python-gettext = "^5.0"
numpy = "^1.24.3"
python-dateutil = "^2.8.2"
aiodocker = "^0.21.0"
telebot = "^0.0.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
